# drivers: eth_mcux: Performance Improvements for MCUX Ethernet Driver