--- a/samples/net/zperf/boards/mimxrt1050_evk.conf
+++ b/samples/net/zperf/boards/mimxrt1050_evk.conf
@@ -1,0 +1,2 @@
+# Note: HW accleration does not support IPV6
+CONFIG_ETH_MCUX_HW_ACCELERATION=y