… samples/net/gptp/boards/mimxrt1050_evk.conf
… samples/net/gptp/boards/mimxrt1060_evk.conf