## Introduction

This PR makes several changes to improve RX/TX performance within the MCUX Ethernet driver. Each commit was tested using an RT1050 and the `samples/net/zperf` sample, to verify that changes were actually improving performance. Performance numbers are given with each commit to help understand which changes had the largest effect.

## Key changes

The primary changes in this PR are the following
- remove the ethernet driver TX thread responsible for reclaiming TX frames, and for adding PTP timestamps. handle this in the ethernet TX isr
- enable hardware checksum calculation for IP packets
- move key network buffers into DTCM to improve data copy performance

This PR also includes a fix for the icmpv4/icmpv6 implementation, to correctly skip IP packet checksums when hardware checksum calculation is enabled

## Performance Results

All results were tested using the `zperf` sample application and iperf version 2.1.5. The "current" revision of Zephyr in these tests was cd7dbe92b1eea7dac07bcac84ef44064d73a940e

target | current TCP TX | current TCP RX | new TCP TX | new TCP RX
-- | -- | -- | -- | --
k64f | 10.3Mbps | 12.83Mbps | 11.9Mbps | 14.1Mbps
mimxrt1060 | 10.4Mbps | 15.8Mbps | 27.0Mbps | 32.4Mbps
mimxrt1050 | 22.4Mbps | 23.1Mbps | 51.1Mbps | 52.4Mbps

## Testing

Beyond performance testing, the gPTP sample `samples/net/gptp` was verified to still work correctly on the `frdm_k64f`, although more rigorous testing is welcomed.